device_proxy: use shared GDR control slab for CI and shutdown
Wire peer×channel rings onto a single [shutdown][ci_0..N) nixlGdrBuffer,
publish CI/reset/shutdown through the helper, and keep the host-mapped
fallback path when GDRCopy is unavailable.

+++ src/core/device_proxy/proxy_worker.h
@@ -32,7 +32,7 @@ class ProxyWorker {
 public:
     ProxyWorker(nixlDeviceProxyBackendAdapter *backend,
                 const nixlProxyMemViewRegistry *proxy_memview_registry,
-                uint32_t *shutdown_word,
+                uint64_t *shutdown_word,
                 nixlProxyChannelState *channels,
                 std::atomic<nixl_proxy_channel_lifecycle_t> *channel_lifecycle,
                 uint32_t peer_capacity,
@@ -82,7 +82,7 @@ private:
 
     nixlDeviceProxyBackendAdapter *backend_ = nullptr;
     const nixlProxyMemViewRegistry *proxy_memview_registry_ = nullptr;
-    uint32_t *shutdown_word_ = nullptr;
+    uint64_t *shutdown_word_ = nullptr;
     nixlProxyChannelState *channels_ = nullptr;
     std::atomic<nixl_proxy_channel_lifecycle_t> *channel_lifecycle_ = nullptr;
     uint32_t peer_capacity_ = 0;
